fix(web): 組織詳細ページが /organizations/:id で開けない問題を修正
TanStack Router のファイルベースルーティングで organizations.$id.tsx が
organizations.tsx の子ルートとして登録されるが、親が Outlet を持たず
OrganizationsPage を component に直接指定していたため、/organizations/:id
にアクセスしても親の一覧ページがそのまま表示されていた。

- organizations.tsx の中身を organizations.index.tsx に分離（一覧ページ）
- organizations.tsx は <Outlet /> を返すだけのレイアウトルートに置き換え
- これにより /organizations は一覧、/organizations/:id は詳細にそれぞれ正しく解決される
- 既存テストのインポートを organizations.index に切り替え

Refs: #88
