+++ apps/web/src/routeTree.gen.ts
@@ -12,6 +12,7 @@ import { Route as rootRouteImport } from './routes/__root'
 import { Route as OrganizationsRouteImport } from './routes/organizations'
 import { Route as LoginRouteImport } from './routes/login'
 import { Route as IndexRouteImport } from './routes/index'
+import { Route as OrganizationsIndexRouteImport } from './routes/organizations.index'
 import { Route as OrganizationsIdRouteImport } from './routes/organizations.$id'
 
 const OrganizationsRoute = OrganizationsRouteImport.update({
@@ -29,6 +30,11 @@ const IndexRoute = IndexRouteImport.update({
   path: '/',
   getParentRoute: () => rootRouteImport,
 } as any)
+const OrganizationsIndexRoute = OrganizationsIndexRouteImport.update({
+  id: '/',
+  path: '/',
+  getParentRoute: () => OrganizationsRoute,
+} as any)
 const OrganizationsIdRoute = OrganizationsIdRouteImport.update({
   id: '/$id',
   path: '/$id',
@@ -40,12 +46,13 @@ export interface FileRoutesByFullPath {
   '/login': typeof LoginRoute
   '/organizations': typeof OrganizationsRouteWithChildren
   '/organizations/$id': typeof OrganizationsIdRoute
+  '/organizations/': typeof OrganizationsIndexRoute
 }
 export interface FileRoutesByTo {
   '/': typeof IndexRoute
   '/login': typeof LoginRoute
-  '/organizations': typeof OrganizationsRouteWithChildren
   '/organizations/$id': typeof OrganizationsIdRoute
+  '/organizations': typeof OrganizationsIndexRoute
 }
 export interface FileRoutesById {
   __root__: typeof rootRouteImport
@@ -53,13 +60,25 @@ export interface FileRoutesById {
   '/login': typeof LoginRoute
   '/organizations': typeof OrganizationsRouteWithChildren
   '/organizations/$id': typeof OrganizationsIdRoute
+  '/organizations/': typeof OrganizationsIndexRoute
 }
 export interface FileRouteTypes {
   fileRoutesByFullPath: FileRoutesByFullPath
-  fullPaths: '/' | '/login' | '/organizations' | '/organizations/$id'
+  fullPaths:
+    | '/'
+    | '/login'
+    | '/organizations'
+    | '/organizations/$id'
+    | '/organizations/'
   fileRoutesByTo: FileRoutesByTo
-  to: '/' | '/login' | '/organizations' | '/organizations/$id'
-  id: '__root__' | '/' | '/login' | '/organizations' | '/organizations/$id'
+  to: '/' | '/login' | '/organizations/$id' | '/organizations'
+  id:
+    | '__root__'
+    | '/'
+    | '/login'
+    | '/organizations'
+    | '/organizations/$id'
+    | '/organizations/'
   fileRoutesById: FileRoutesById
 }
 export interface RootRouteChildren {
@@ -91,6 +110,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof IndexRouteImport
       parentRoute: typeof rootRouteImport
     }
+    '/organizations/': {
+      id: '/organizations/'
+      path: '/'
+      fullPath: '/organizations/'
+      preLoaderRoute: typeof OrganizationsIndexRouteImport
+      parentRoute: typeof OrganizationsRoute
+    }
     '/organizations/$id': {
       id: '/organizations/$id'
       path: '/$id'
@@ -103,10 +129,12 @@ declare module '@tanstack/react-router' {
 
 interface OrganizationsRouteChildren {
   OrganizationsIdRoute: typeof OrganizationsIdRoute
+  OrganizationsIndexRoute: typeof OrganizationsIndexRoute
 }
 
 const OrganizationsRouteChildren: OrganizationsRouteChildren = {
   OrganizationsIdRoute: OrganizationsIdRoute,
+  OrganizationsIndexRoute: OrganizationsIndexRoute,
 }
 
 const OrganizationsRouteWithChildren = OrganizationsRoute._addFileChildren(
