+++ apps/web/src/test/organizations.test.tsx
@@ -1,7 +1,7 @@
 import { screen, waitFor, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import { OrganizationsPage } from "../routes/organizations";
+import { OrganizationsPage } from "../routes/organizations.index";
 import { renderWithQuery } from "./test-utils";
 
 // hono/client api モジュールをモック
